Separate thread for linux-network-watcher
